# Decrease allocations